elevaite/creative_dashboard_backend/pyproject.toml
[project]
name = "creative-dashboard-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "vishnukk929", email = "vishnu.krishnan@iopex.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain  (==0.3.20)",
    "langchain-openai (>=0.3.9,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)"
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain = "==0.3.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
